// packages/unified/src/remark/utils.d.ts
export type LinkFilter = {
    loc?: Partial<DNoteLoc>;
};
export type ParseLinkV2Resp = {
    alias?: string;
    value: string;
    anchorHeader?: string;
    vaultName?: string;
    sameFile: false;
} | {
    alias?: string;
    value?: string;
    anchorHeader: string;
    vaultName?: string;
    sameFile: true;
};
export declare function hashTag2WikiLinkNoteV4(hashtag: HashTag): WikiLinkNoteV4;
export declare function userTag2WikiLinkNoteV4(userTag: UserTag): WikiLinkNoteV4;
export declare function frontmatterTag2WikiLinkNoteV4(tag: string, useHashSymbol?: boolean): WikiLinkNoteV4;
export declare class LinkUtils {
    static astType2DLinkType(type: DendronASTTypes): DLink["type"];
    static dlink2DNoteLink(link: DLink): DNoteLink;
    /**
     * Get links from note body while maintaining existing backlinks
     */
    static findLinks({ note, engine, type, config, filter, }: {
        note: NoteProps;
        engine: ReducedDEngine;
        config: DendronConfig;
        filter?: LinkFilter;
        type: "regular" | "candidate";
    }): Promise<DLink[]>;
    /**
     * Get all links from the note body
     * Currently, just look for wiki links
     *
     * @param opts.filter - {type, loc
     *
     * - type: filter by {@link DendronASTTypes}
     * - loc: filter by {@link DLoc}
     */
    static findLinksFromBody({ note, config, filter, }: {
        note: NoteProps;
        config: DendronConfig;
        filter?: LinkFilter;
    }): DLink[];
    static findHashTags({ links }: {
        links: DLink[];
    }): DLink[];
    static isAlias(link: string): boolean;
    static hasFilter(link: string): boolean;
    static parseAliasLink(link: string): {
        alias: string | undefined;
        value: string;
    };
    /** Either value or anchorHeader will always be present if the function did not
     *  return null. A missing value means that the file containing this link is
     *  the value.
     *
     *  if `explicitAlias` is false, non-existent alias will be
     *  implicitly assumed to be the value of the link.
     */
    static parseLinkV2(opts: {
        linkString: string;
        explicitAlias?: boolean;
    }): ParseLinkV2Resp | null;
    static getNotesFromWikiLinks(opts: {
        activeNote: DNodeProps;
        wikiLinks: ParseLinkV2Resp[];
        engine: DEngineClient;
    }): Promise<NoteProps[]>;
    static parseLink(linkMatch: string): WikiLinkProps;
    static parseNoteRefRaw(ref: string): DNoteRefLinkRaw;
    static parseNoteRef(ref: string): DNoteRefLink;
    static renderNoteLink({ link, dest, }: {
        link: DNoteLink;
        dest: DendronASTDest;
    }): string;
    /**
     * Given a note, updates old link to new link.
     * If you are calling this on a note multiple times,
     * You need to start from the last link in the body
     * Because each updateLink call will shift the location of
     * every element in the note body that comes after
     * the link you update
     *
     * @param note the note that contains link to update
     * @param oldLink the old link that needs to be updated
     * @param newLink new link
     * @returns
     */
    static updateLink({ note, oldLink, newLink, }: {
        note: NoteProps;
        oldLink: DNoteLink;
        newLink: DNoteLink;
    }): string;
    static isHashtagLink(link: DNoteLoc): link is DNoteLoc & {
        alias: string;
    };
    static isUserTagLink(link: DNoteLoc): link is DNoteLoc & {
        alias: string;
    };
    static findLinkCandidates({ note, engine, config, }: {
        note: NoteProps;
        engine: ReducedDEngine;
        config: DendronConfig;
    }): Promise<DLink[]>;
    /**
     *  Use this version during engine initialization on the engine side, where
     *  the entire set of noteDicts is available. This version uses local data so
     *  it's much faster.
     * @param param0
     * @returns
     */
    static findLinkCandidatesSync({ note, noteDicts, config, }: {
        note: NoteProps;
        noteDicts: NoteDicts;
        config: DendronConfig;
    }): DLink[];
    static hasVaultPrefix(link: DLink): boolean;
    /**
     * Given a source string, extract all wikilinks within the source.
     *
     * @param source string to extract wikilinks from
     */
    static extractWikiLinks(source: string): ParseLinkV2Resp[];
    /**
     * Given an array of links that need to be updated,
     * and a note that contains the links,
     * update all links in given array so that they point to
     * the destination note.
     *
     * returns note change entries
     */
    static updateLinksInNote(opts: {
        linksToUpdate: DNoteLink[];
        note: NoteProps;
        destNote: NoteProps;
        engine: DEngineClient;
    }): Promise<import("@dendronhq/common-all").WriteNoteResp>;
}